searchMusic/requirements.txt
grpcio
grpcio-tools
ytmusicapi
protobuf